pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Fotis", email = "fotiskapotos@gmail.com" }
]
requires-python = ">=3.9"
dependencies = [
    "gymnasium>=0.29.0",
    "matplotlib>=3.9.4",
    "pillow>=11.3.0",
    "text-flappy-bird-gym",
    "tqdm>=4.67.3",
    "wandb[media]>=0.25.1",
]

[project.scripts]
project = "project:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
text-flappy-bird-gym = { git = "https://gitlab-research.centralesupelec.fr/stergios.christodoulidis/text-flappy-bird-gym.git" }

[dependency-groups]
dev = [
    "ipykernel>=6.31.0",
]
